requirements.txt
celery==5.4.0
python-dotenv==1.0.1
pytubefix==6.13.1